+++ .gitignore
@@ -29,3 +29,4 @@ ormconfig.json
 
 # Build
 dist/
+.nuxt/
