--- nuxt.config.js
@@ -1,32 +1,0 @@
-module.exports = {
-	mode: 'spa',
-	head: {
-		title: 'Codes',
-		meta: [
-			{ charset: 'utf-8' },
-			{ name: 'viewport', content: 'width=device-width, initial-scale=1' },
-			{ hid: 'description', name: 'description', content: 'Codes website for multipurpose functions' }
-		],
-		link: [
-			{ rel: 'icon', type: 'image/x-icon', href: '/favicon.png' }
-		]
-	},
-	srcDir: `${process.cwd()}/src/dashboard`,
-	// rootDir: `${process.cwd()}/src/dashboard`,
-	buildDir: `${process.cwd()}/dist/dashboard`,
-	loading: { color: '#00f' },
-	modules: [
-		// Doc: https://bootstrap-vue.js.org
-		'bootstrap-vue/nuxt',
-		// Doc: https://axios.nuxtjs.org/usage
-		'@nuxtjs/axios'
-	],
-	/*
-	** Axios module configuration
-	** See https://axios.nuxtjs.org/options
-	*/
-	axios: {
-		baseURL: '/',
-		progress: true
-	},
-};
