+++ package.json
@@ -1,54 +1,6 @@
 {
-  "private": true,
-  "name": "codes-support",
-  "version": "1.0.0",
-  "main": "dist/index.js",
-  "author": "Seif Mansour",
-  "license": "UNLICENSED",
-  "scripts": {
-    "start": "cross-env NODE_ENV=production node .",
-    "dev": "cross-env NODE_ENV=development node .",
-    "build": "nuxt build && tsc",
-    "watch": "tsc -w",
-    "lint": "eslint src --ext .ts",
-    "fix": "eslint src --ext .ts --fix"
-  },
-  "dependencies": {
-    "@nuxtjs/axios": "^5.8.0",
-    "bootstrap": "^4.3.1",
-    "bootstrap-vue": "^2.0.4",
-    "common-tags": "^1.8.0",
-    "cross-env": "^6.0.3",
-    "discord-akairo": "discord-akairo/discord-akairo",
-    "discord.js": "discordjs/discord.js",
-    "express": "^4.17.1",
-    "mongodb": "^3.3.3",
-    "morgan": "^1.9.1",
-    "node-fetch": "^2.6.0",
-    "nuxt": "^2.10.2",
-    "shortid": "^2.2.15",
-    "typeorm": "^0.2.20",
-    "winston": "^3.2.1",
-    "winston-daily-rotate-file": "^4.2.1"
-  },
-  "devDependencies": {
-    "@types/common-tags": "^1.8.0",
-    "@types/dotenv": "^8.2.0",
-    "@types/express": "^4.17.2",
-    "@types/morgan": "^1.7.37",
-    "@types/node-fetch": "^2.5.3",
-    "@types/shortid": "^0.0.29",
-    "@types/ws": "^6.0.3",
-    "@typescript-eslint/eslint-plugin": "^2.6.0",
-    "@typescript-eslint/parser": "^2.6.0",
-    "eslint": "^6.6.0",
-    "eslint-config-marine": "^5.3.2",
-    "typescript": "^3.6.4"
-  },
-  "eslintConfig": {
-    "extends": "marine/node",
-    "rules": {
-      "@typescript-eslint/explicit-member-accessibility": "off"
-    }
-  }
+	"private": true,
+	"workspaces": [
+		"packages/*"
+	]
 }
