+++ packages/bot/src/util/env.ts
@@ -1,0 +1,4 @@
+import { config } from 'dotenv';
+import { join } from 'path';
+
+config({ path: join(process.cwd(), '..', '..', '.env') });
