+++ packages/dashboard/src/components/Navbar.vue
