+++ packages/dashboard/src/static/favicon.png
