+++ packages/server/package.json
@@ -1,0 +1,43 @@
+{
+	"name": "@codes/server",
+	"version": "1.0.0",
+	"main": "dist/index.js",
+	"author": "GamesProSeif",
+	"license": "UNLICENSED",
+	"scripts": {
+		"start": "cross-env NODE_ENV=production node .",
+		"dev": "cross-env NODE_ENV=development node .",
+		"build": "tsc",
+		"build:watch": "tsc -w",
+		"lint": "eslint src --ext .ts",
+		"lint:fix": "eslint src --ext .ts --fix"
+	},
+	"dependencies": {
+		"cross-env": "^6.0.3",
+		"dotenv": "^8.2.0",
+		"express": "^4.17.1",
+		"mongodb": "^3.3.3",
+		"morgan": "^1.9.1",
+		"shortid": "^2.2.15",
+		"typeorm": "^0.2.20",
+		"winston": "^3.2.1",
+		"winston-daily-rotate-file": "^4.2.1"
+	},
+	"devDependencies": {
+		"@types/dotenv": "^8.2.0",
+		"@types/express": "^4.17.2",
+		"@types/morgan": "^1.7.37",
+		"@types/shortid": "^0.0.29",
+		"@typescript-eslint/eslint-plugin": "^2.6.0",
+		"@typescript-eslint/parser": "^2.6.0",
+		"eslint": "^6.6.0",
+		"eslint-config-marine": "^5.3.2",
+		"typescript": "^3.6.4"
+	},
+	"eslintConfig": {
+		"extends": "marine/node",
+		"rules": {
+			"@typescript-eslint/explicit-member-accessibility": "off"
+		}
+	}
+}
