--- src/dashboard/README.md
@@ -1,3 +1,0 @@
-# Dashboard
-
-Nuxt files here
