--- src/util/env.ts
@@ -1,4 +1,0 @@
-import { config } from 'dotenv';
-import { join } from 'path';
-
-config({ path: join(process.cwd(), '.env') });
